% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetReader.R
\name{datasetReader}
\alias{datasetReader}
\title{Read Dataset File into Memory.}
\usage{
datasetReader(directory, testDirectory, selectedFeats = c(),
  classCol = "class", preProcessF = "N", featuresToPreProcess = c(),
  nComp = NA, missingVal = c("NA", "?", " "), missingOpr = 0)
}
\arguments{
\item{directory}{String of the directory to the file containing the training dataset.}

\item{testDirectory}{String of the directory to the file containing the testing dataset.}

\item{selectedFeats}{Vector of numbers of features columns to include from the training set and ignore the rest of columns - In case of empty vector, this means to include all features in the dataset file (default = c()).}

\item{classCol}{String of the name of the class label column in the dataset (default = 'class').}

\item{preProcessF}{string containing the name of the preprocessing algorithm (default = 'N' --> no preprocessing):
\itemize{
\item "boxcox" - apply a Box–Cox transform and values must be non-zero and positive in all features,
\item "yeo-Johnson" - apply a Yeo-Johnson transform, like a BoxCox, but values can be negative,
\item "zv" - remove attributes with a zero variance (all the same value),
\item "center" - subtract mean from values,
\item "scale" - divide values by standard deviation,
\item "standardize" - perform both centering and scaling,
\item "normalize" - normalize values,
\item "pca" - transform data to the principal components,
\item "ica" - transform data to the independent components.
}}

\item{featuresToPreProcess}{Vector of number of features to perform the feature preprocessing on - In case of empty vector, this means to include all features in the dataset file (default = c()) - This vector should be a subset of \code{selectedFeats}.}

\item{nComp}{Integer of Number of components needed if either "pca" or "ica" feature preprocessors are needed.}

\item{missingVal}{Vector of strings representing the missing values in dataset (default: c('NA', '?', ' ')).}

\item{missingOpr}{Boolean variable represents either delete instances with missing values or apply imputation using "MICE" library which helps you imputing missing values with plausible data values that are drawn from a distribution specifically designed for each missing datapoint- (default = 0 --> delete instances).}
}
\value{
List of the TrainingSet \code{Train} and TestingSet \code{Test}.
}
\description{
Read the file of the training and testing dataset, and perform preprocessing and data cleaning if necessary.
}
\examples{
\dontrun{
dataset <- datasetReader('/Datasets/irisTrain.csv', '/Datasets/irisTest.csv')
}
}
