% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClassifier.R
\name{runClassifier}
\alias{runClassifier}
\title{Fit a classifier model.}
\usage{
runClassifier(trainingSet, validationSet, params, classifierAlgorithm,
  interp = 0)
}
\arguments{
\item{trainingSet}{Dataframe of the training set.}

\item{validationSet}{Dataframe of the validation Set.}

\item{params}{A string of parameter configuration values for the current classifier to be tuned (parameters are separated by #).}

\item{classifierAlgorithm}{String of the name of classifier algorithm used now.}

\item{interp}{Boolean representing if interpretability is required or not (Default = 0).}
}
\value{

}
\description{
Run the classifier on a training set and measure performance on a validation set.
}
\examples{

}
